--- a/docs/database-optimization.md
+++ b/docs/database-optimization.md
@@ -1,0 +1,65 @@
+# Database Optimization Strategy
+
+## Overview
+This document outlines the database optimization strategies implemented in the application to improve query performance and overall system efficiency.
+
+## Implemented Optimizations
+
+### 1. Indexing Strategy
+- Added indexes on frequently queried columns:
+  - `email` (unique index)
+  - `role` (for role-based queries)
+  - `status` (for filtering active/inactive users)
+  - `createdAt` (for date-based queries and sorting)
+
+### 2. Query Optimizations
+- **Pagination**: Implemented in `findAll` queries to limit result sets
+- **Selective Loading**: Using `select` option to fetch only required fields
+- **Relation Loading**: Optimized with `relations` parameter
+- **Query Caching**: Implemented with 1-minute cache duration
+- **Soft Delete**: Replaced hard deletes with soft deletes
+
+### 3. Connection Pool Optimization
+- Increased max pool size to 20
+- Set minimum pool size to 5
+- Configured idle timeout to 30 seconds
+- Implemented retry mechanism (5 attempts with 3-second delay)
+
+### 4. Performance Monitoring
+- Implemented `DatabaseMonitoringService` for:
+  - Slow query detection (threshold: 1000ms)
+  - Query logging in development
+  - Performance metrics collection
+
+### 5. Error Handling
+- Implemented specific error types
+- Added comprehensive try-catch blocks
+- Improved error messages for better debugging
+
+## Performance Tests
+- Added performance test suite in `users.service.spec.ts`
+- Tests cover:
+  - Pagination performance
+  - Relation loading efficiency
+  - Concurrent request handling
+
+## Monitoring and Metrics
+- Query execution time tracking
+- Slow query detection and logging
+- Connection pool usage monitoring
+- Error rate tracking
+
+## Best Practices
+1. Always use pagination for list queries
+2. Select only required fields
+3. Use appropriate indexes
+4. Implement proper error handling
+5. Monitor query performance
+6. Use connection pooling effectively
+
+## Future Improvements
+1. Implement read replicas for heavy read operations
+2. Add more comprehensive performance metrics
+3. Implement query result caching for frequently accessed data
+4. Regular database maintenance scheduling
+5. Consider implementing database sharding for large datasets